replace ORB.resolve() with ORB.stringToObject("corbaname:...")
test/giop.spec.ts now works using this line

  orb.stringToObject("corbaname::192.168.1.10#TestService")

even with nameservice and object on different ORBs. but it's a ugly
disgusting hack for now.

next step will be to add a minimal NameService to corba.js and replace
ORB.bind().

after that we should have most of GIOP in place to get the whole
testsuite running again.